Admin evetns

+++ src/BLL/EventService.java
@@ -1,7 +1,10 @@
 package BLL;
 
 import BE.Event;
+import BE.EventCoordinator;
+import BE.EventCoordinatorEvent;
 import DAL.AddEvent;
+import DAL.DeleteECEvent;
 import DAL.DeleteEvent;
 import DAL.DeleteUserEvent;
 
@@ -12,19 +15,37 @@ public class EventService {
     private final DeleteEvent deleteEvent = new DeleteEvent();
     private final AddEvent addEvent = new AddEvent();
     private final DeleteUserEvent deleteUserEvent = new DeleteUserEvent();
+    private final DeleteECEvent deleteECEvent = new DeleteECEvent();
 
-    public void addEvent(Event event){
-        single.addEventSingle(addEvent.newEvent(event));
+    public Event addEvent(Event event){
+        return single.addEventSingle(addEvent.newEvent(event));
     }
 
     public void deleteEvent(Event event){
         if(deleteUserEvent.deleteUserEvent(event)){
             single.deleteUserEvent(event);
-            if(deleteEvent.deleteEvent(event)){
-                System.out.println("test");
-                single.deleteEventSingle(event);
+            if (deleteECEvent.deleteECEvent(event)){
+                single.deleteECEvent(event);
+                if(deleteEvent.deleteEvent(event)){
+                    System.out.println("test");
+                    single.deleteEventSingle(event);
+                }
             }
         }
+    }
+
+    public ArrayList<Event> getEventsForEC(EventCoordinator ec) {
+        ArrayList<Event> placeholder = new ArrayList<>();
+        for (EventCoordinatorEvent e:single.getECEvents()) {
+            if (e.getECID() == ec.getId()){
+                for (Event event:single.getEvents()) {
+                    if (e.getEventID()==event.getId()){
+                        placeholder.add(event);
+                    }
+                }
+            }
+        }
+        return placeholder;
     }
 
     public ArrayList<Event> getEvents (){
